Change scoring-fn to provide more "granular" scoring
Previously, if *any* event for a guest was double-scheduled or outside of their availability, it would return a score of 999; even if it was 1, 10 or 100 events. 
Thus, the annealing could only "make progress" if it was able to generate a new state that completely moved a guest's schedule from "faulty" to "not-faulty".
Now, the guests score changes "linearly" with the number of faults, allowing the annealing to make smaller progress.